Refactor: extract onboard Group C lifecycle into DeviceRunnerBase
Move the byte-identical lifecycle code from a2a3 and a5 onboard
DeviceRunners into the shared DeviceRunnerBase (-354 net lines):

- Methods: create_thread, attach_current_thread,
  configure_aicore_op_timeout, ensure_device_initialized,
  ensure_binaries_loaded, query_max_block_dim, validate_block_dim,
  print_handshake_results, set_executors, set_dispatcher_binary,
  device_id(), last_device_wall_ns().
- State: device_id_, block_dim_, cores_per_blockdim_, worker_count_,
  aicpu_so_binary_, aicore_kernel_binary_, aicore_bin_handle_,
  dispatcher_so_binary_, load_aicpu_op_, stream_aicpu_, stream_aicore_,
  kernel_args_, device_wall_dev_ptr_, device_wall_ns_, device_args_,
  binaries_loaded_.

setup_static_arena, finalize, run, the kernel launch path, the
prepared-callable registry, profiling-init methods, the comm/HCCL
plumbing (a2a3 only), and dep_gen (a2a3 only) stay arch-specific —
each has a real semantic divergence between arches.

Verbatim code moves; no behavioral changes. nm -D confirms both
arches' libhost_runtime.so export the lifecycle methods on
DeviceRunnerBase and keep the arch-specific symbols on DeviceRunner.
Local a2a3 onboard smoke (dummy_task, alternating_matmul_add,
prepared_callable) passes 8/8 in 15s.
