+++ src/a2a3/platform/onboard/host/device_runner.h
@@ -92,16 +92,13 @@ public:
      * Returns 0 on success, -1 on failure.
      *
      * `allocate_tensor`, `free_tensor`, `copy_to_device`, `copy_from_device`,
-     * and `acquire_pooled_{gm_heap,gm_sm,runtime_arena}` are inherited from
+     * `acquire_pooled_{gm_heap,gm_sm,runtime_arena}`, `create_thread`,
+     * `attach_current_thread`, `ensure_device_initialized`,
+     * `print_handshake_results`, `set_executors`, `set_dispatcher_binary`,
+     * `device_id`, and `last_device_wall_ns` are inherited from
      * `DeviceRunnerBase`.
      */
     int setup_static_arena(size_t gm_heap_size, size_t gm_sm_size, size_t runtime_arena_size);
-
-    /**
-     * Create a thread bound to this device.
-     * The thread calls rtSetDevice(device_id) on entry.
-     */
-    std::thread create_thread(std::function<void()> fn);
 
     /**
      * Execute a runtime
@@ -129,33 +126,6 @@ public:
     int run(Runtime &runtime, int block_dim, int launch_aicpu_num = 1);
 
     /**
-     * Take ownership of the AICPU + AICore executor binaries. Called once
-     * by simpler_init at ChipWorker::init time; subsequent run() invocations
-     * read from `aicpu_so_binary_` / `aicore_kernel_binary_`.
-     */
-    void set_executors(std::vector<uint8_t> aicpu_so_binary, std::vector<uint8_t> aicore_kernel_binary) {
-        aicpu_so_binary_ = std::move(aicpu_so_binary);
-        aicore_kernel_binary_ = std::move(aicore_kernel_binary);
-    }
-
-    /**
-     * Take ownership of the dispatcher SO bytes. Called by simpler_init when
-     * the caller provided a dispatcher path; the eager
-     * ensure_device_initialized() in simpler_init hands the buffer to
-     * LoadAicpuOp::BootstrapDispatcher at init time. Leaving this unset
-     * (empty buffer) makes ensure_binaries_loaded() fail with a clear
-     * message — callers that drive _ChipWorker.init directly without a
-     * dispatcher path get a deterministic error at simpler_init time rather
-     * than a confusing dladdr-derived path.
-     */
-    void set_dispatcher_binary(std::vector<uint8_t> dispatcher_so_binary) {
-        dispatcher_so_binary_ = std::move(dispatcher_so_binary);
-    }
-
-    /** The device id captured by simpler_init's attach_current_thread call. */
-    int device_id() const { return device_id_; }
-
-    /**
      * Enablement setters for the three diagnostics sub-features. Called by
      * the c_api entry point before run(); downstream run() paths read the
      * corresponding `enable_*_` members directly. Moved off the generic
@@ -177,22 +147,6 @@ public:
     // is enabled; CallConfig::validate() enforces this contract upstream.
     void set_output_prefix(const char *prefix) { output_prefix_ = (prefix != nullptr) ? prefix : ""; }
     const std::string &output_prefix() const { return output_prefix_; }
-
-    /**
-     * Device-side wall (ns) from the most recently completed run, written
-     * by the platform AICPU entry (onboard: kernel.cpp; sim: lambda in
-     * run()). Returns 0 before any run completes. Independent of any
-     * profiling / swimlane subsystem.
-     */
-    uint64_t last_device_wall_ns() const { return device_wall_ns_; }
-
-    /**
-     * Print handshake results from device
-     *
-     * Copies handshake buffers from device and prints their status.
-     * Must be called after run() and before finalize().
-     */
-    void print_handshake_results();
 
     /**
      * Cleanup all resources
@@ -254,17 +208,6 @@ public:
     uint64_t upload_chip_callable_buffer(const ChipCallable *callable);
 
     /**
-     * Attach the current host thread to the target device.
-     *
-     * This is required before host-side runtime initialization may allocate or
-     * free device memory on the current thread. No streams are created here.
-     *
-     * @param device_id  Device ID (0-15)
-     * @return 0 on success, error code on failure
-     */
-    int attach_current_thread(int device_id);
-
-    /**
      * Make the ACL context ready on the current thread.
      *
      * Calls aclInit() once per process (subsequent calls are idempotent and
@@ -299,29 +242,6 @@ public:
      * @return 0 on success, error code on failure.
      */
     int destroy_comm_stream(void *stream);
-
-    /**
-     * One-shot device initialization. Performs, in order:
-     *   1. rtSetDevice + rtStreamCreate for AICPU and AICore streams. Streams
-     *      live for the DeviceRunner's lifetime and are destroyed in finalize.
-     *   2. Bundles dispatcher SO bytes + inner AICPU kernel SO bytes through
-     *      `LoadAicpuOp::BootstrapDispatcher` so the inner SO is written to
-     *      the device-side preinstall path.
-     *   3. Registers the inner SO via `LoadAicpuOp::Init`
-     *      (`rtsBinaryLoadFromFile` + `rtsFuncGetByName`) and caches the
-     *      resulting per-symbol `rtFuncHandle` for per-task `rtsLaunchCpuKernel`.
-     *   4. H2D-copies the (zeroed) per-task DeviceArgs struct via
-     *      `kernel_args_.init_device_args`. device_args_.aicpu_so_bin/len
-     *      stay 0 — no consumer reads them on the per-task path.
-     *
-     * Called once from `simpler_init` after the executor + dispatcher bytes are
-     * cached on the runner. Idempotent: subsequent calls short-circuit on
-     * binaries_loaded_. Reads device_id_ recorded by attach_current_thread.
-     *
-     * @return 0 on success, error code on failure (e.g. dispatcher SO bytes
-     *         not provided, CANN stream create / register failures).
-     */
-    int ensure_device_initialized();
 
     /**
      * Stage a per-callable_id orchestration SO into device memory and remember
@@ -424,75 +344,18 @@ public:
     size_t host_dlopen_count() const { return host_dlopen_total_; }
 
 private:
-    // Internal state. device_id_ is set once in attach_current_thread() (called
-    // from simpler_init during ChipWorker::init) and read on every subsequent
-    // op. All ChipWorker callers run on the same thread that called init, so
-    // plain int + the init→user happens-before edge is sufficient.
-    int device_id_{-1};
-    int block_dim_{0};
-    int cores_per_blockdim_{PLATFORM_CORES_PER_BLOCKDIM};
-    int worker_count_{0};  // Stored for print_handshake_results in destructor
-    // Executor binaries — populated once via set_executors() during
-    // simpler_init. aicore_kernel_binary_ is consumed once by
-    // launch_aicore_kernel() (rtRegisterAllKernel returns aicore_bin_handle_,
-    // cached and reused on every subsequent launch). Caching is required:
-    // CANN has no public rtUnregisterAllKernel, so re-registering on every
-    // run would pin another device-side copy of the ELF and quickly exhaust
-    // HBM (manifested in CI as 207001 at rtKernelLaunchWithHandleV2 with a
-    // 507899 cascade at rtStreamCreate). aicpu_so_binary_ is released by
-    // ensure_binaries_loaded() after bootstrap; bootstrap is the only
-    // consumer and per-task launches go through the cached rtFuncHandle on
-    // LoadAicpuOp, not the host bytes.
-    std::vector<uint8_t> aicpu_so_binary_;
-    std::vector<uint8_t> aicore_kernel_binary_;
-    // AICore kernel handle from rtRegisterAllKernel — lazily populated by
-    // launch_aicore_kernel() and reused across all runs. nullptr means not
-    // yet registered. Reset to nullptr in finalize(); CANN releases the
-    // device-side state implicitly when the device context tears down.
-    void *aicore_bin_handle_{nullptr};
-    // Dispatcher SO bytes — populated once via set_dispatcher_binary() during
-    // simpler_init. Consumed exclusively by BootstrapDispatcher on the first
-    // run() and released by ensure_binaries_loaded() right after. Empty buffer
-    // is permitted at init time (callers that drive ChipWorker.init without a
-    // dispatcher path); ensure_binaries_loaded() then fails fast with a clear
-    // message if/when bootstrap is actually attempted.
-    std::vector<uint8_t> dispatcher_so_binary_;
-
-    // AICPU op loader — handles dispatcher bootstrap and per-task launches.
-    host::LoadAicpuOp load_aicpu_op_;
-
-    // `mem_alloc_`, `gm_heap_arena_`, `gm_sm_arena_`, `runtime_arena_pool_`,
-    // and the alloc/free trampolines are inherited from `DeviceRunnerBase`.
+    // Most lifecycle state (device_id_, block_dim_, cores_per_blockdim_,
+    // worker_count_, executor + dispatcher bytes, aicore_bin_handle_,
+    // load_aicpu_op_, mem_alloc_, the three DeviceArenas, persistent
+    // AICPU/AICore streams, kernel_args_, device_wall_*, device_args_,
+    // binaries_loaded_) is inherited from `DeviceRunnerBase`.
     //
-    // Released explicitly in finalize() before mem_alloc_.finalize() so the
-    // underlying buffers do not get freed twice. `runtime_arena_pool_` stays
-    // unreserved when setup_static_arena was invoked with
-    // runtime_arena_size == 0 (hbg path).
-    //
-    // Cached sizes for setup_static_arena's "fits" check — avoids re-allocating
-    // the same buffer when a later worker init asks for an equal-or-smaller
-    // layout on an already-committed arena.
+    // Arena cached sizes for setup_static_arena's "fits" check — avoids
+    // re-allocating the same buffer when a later worker init asks for an
+    // equal-or-smaller layout on an already-committed arena.
     size_t cached_gm_heap_size_{0};
     size_t cached_gm_sm_size_{0};
     size_t cached_runtime_arena_size_{0};
-
-    // Device resources
-    rtStream_t stream_aicpu_{nullptr};
-    rtStream_t stream_aicore_{nullptr};
-    KernelArgsHelper kernel_args_;
-
-    // Platform-level device wall buffer: 8-byte device-resident slot whose
-    // address rides on KernelArgs.device_wall_data_base. AICPU writes the
-    // run wall (ns) through that pointer; this DeviceRunner pulls it back
-    // via copy_from_device after stream sync and caches it for
-    // last_device_wall_ns(). Allocated once at simpler_init, freed in
-    // finalize.
-    void *device_wall_dev_ptr_{nullptr};
-    uint64_t device_wall_ns_{0};
-    DeviceArgs device_args_;
-
-    // Kernel binary management
-    bool binaries_loaded_{false};  // true after AICPU SO loaded
 
     // Chip-callable buffer pool. Keyed by FNV-1a 64-bit content hash of the
     // ChipCallable bytes. Each entry owns one device GM allocation holding
@@ -562,39 +425,9 @@ private:
     // dep_gen collector — captures orchestrator submit_task inputs for offline replay
     DepGenCollector dep_gen_collector_;
 
-    /**
-     * Query the maximum block_dim the stream can host.
-     *
-     * Uses aclrtGetStreamResLimit(CUBE_CORE / VECTOR_CORE) and returns
-     * min(cube / AIC_PER_BLOCKDIM, vector / AIV_PER_BLOCKDIM). Falls back to
-     * the static PLATFORM_MAX_BLOCKDIM cap when the query is unavailable or
-     * reports no cores. Used both to validate explicit block_dim values and
-     * to resolve the CallConfig "auto" sentinel (block_dim == 0).
-     *
-     * If non-null, `out_cube` / `out_vector` receive the raw ACL limits when
-     * the query succeeded, or 0 when it failed. Callers use this to
-     * distinguish the ACL-unavailable fallback path from the success path in
-     * error logs.
-     */
-    int query_max_block_dim(rtStream_t stream, uint32_t *out_cube = nullptr, uint32_t *out_vector = nullptr);
-
-    /**
-     * Validate block_dim against the stream's CUBE/VECTOR core limits
-     * (via query_max_block_dim). Returns 0 if block_dim fits, -1 otherwise
-     * (or if block_dim < 1).
-     */
-    int validate_block_dim(rtStream_t stream, int block_dim);
-
-    /**
-     * Load AICPU SO and initialize device args
-     *
-     * Called from ensure_device_initialized() after the persistent streams
-     * are created. Reads aicpu_so_binary_ / aicore_kernel_binary_ off the
-     * runner.
-     *
-     * @return 0 on success, error code on failure
-     */
-    int ensure_binaries_loaded();
+    // `query_max_block_dim`, `validate_block_dim`, `ensure_binaries_loaded`,
+    // and `configure_aicore_op_timeout` are inherited (protected) from
+    // `DeviceRunnerBase`.
 
     /**
      * Stamp `runtime.{dev_orch_so_addr_, dev_orch_so_size_}` from the
@@ -608,16 +441,6 @@ private:
      * @return 0 on success, non-zero on failure.
      */
     int prepare_orch_so(Runtime &runtime);
-
-    /**
-     * Configure STARS op execution timeout (once per DeviceRunner lifetime).
-     *
-     * Called on first device attach to set the hardware-level AICore op
-     * execution timeout via aclrtSetOpExecuteTimeOutV2.  The actual
-     * timeout may differ from the requested value due to hardware timer
-     * granularity.
-     */
-    void configure_aicore_op_timeout();
 
     /**
      * Initialize performance profiling shared memory
