+++ src/a5/platform/onboard/host/device_runner.cpp
@@ -110,213 +110,10 @@ int DeviceRunner::setup_static_arena(size_t gm_heap_size, size_t gm_sm_size, siz
     return 0;
 }
 
-std::thread DeviceRunner::create_thread(std::function<void()> fn) {
-    int dev_id = device_id_;
-    return std::thread([dev_id, fn = std::move(fn)]() {
-        rtSetDevice(dev_id);
-        fn();
-    });
-}
-
-int DeviceRunner::ensure_device_initialized() {
-    // Attach the current thread to the device (device_id_ was set in
-    // attach_current_thread() during simpler_init) and create the persistent
-    // AICPU/AICore streams. Streams live for the DeviceRunner's lifetime and
-    // are destroyed in finalize().
-    int rc = attach_current_thread(device_id_);
-    if (rc != 0) {
-        return rc;
-    }
-
-    bool aicpu_created_here = false;
-    bool aicore_created_here = false;
-    if (stream_aicpu_ == nullptr) {
-        rc = rtStreamCreate(&stream_aicpu_, 0);
-        if (rc != 0) {
-            LOG_ERROR("rtStreamCreate (AICPU) failed: %d", rc);
-            return rc;
-        }
-        aicpu_created_here = true;
-    }
-    if (stream_aicore_ == nullptr) {
-        rc = rtStreamCreate(&stream_aicore_, 0);
-        if (rc != 0) {
-            LOG_ERROR("rtStreamCreate (AICore) failed: %d", rc);
-            // Roll back only the AICPU stream we just created, not a
-            // pre-existing persistent one.
-            if (aicpu_created_here) {
-                rtStreamDestroy(stream_aicpu_);
-                stream_aicpu_ = nullptr;
-            }
-            return rc;
-        }
-        aicore_created_here = true;
-    }
-    if (aicpu_created_here || aicore_created_here) {
-        LOG_INFO_V0("DeviceRunner: device=%d set, streams created", device_id_);
-    }
-
-    return ensure_binaries_loaded();
-}
-
-int DeviceRunner::attach_current_thread(int device_id) {
-    if (device_id < 0) {
-        LOG_ERROR("Invalid device_id: %d", device_id);
-        return -1;
-    }
-    if (device_id_ != -1 && device_id_ != device_id) {
-        LOG_ERROR(
-            "DeviceRunner already initialized on device %d; reset/finalize before switching to device %d", device_id_,
-            device_id
-        );
-        return -1;
-    }
-
-    // CANN device context is per-thread, so every caller must attach explicitly.
-    int rc = rtSetDevice(device_id);
-    if (rc != 0) {
-        LOG_ERROR("rtSetDevice(%d) failed: %d", device_id, rc);
-        return rc;
-    }
-
-    if (device_id_ == -1) {
-        configure_aicore_op_timeout();
-    }
-
-    device_id_ = device_id;
-    return 0;
-}
-
-void DeviceRunner::configure_aicore_op_timeout() {
-    uint64_t actual_timeout = 0;
-    int rc = aclrtSetOpExecuteTimeOutV2(PLATFORM_OP_EXECUTE_TIMEOUT_US, &actual_timeout);
-    if (rc != 0) {
-        LOG_ERROR(
-            "aclrtSetOpExecuteTimeOutV2(%llu us) failed: %d", (unsigned long long)PLATFORM_OP_EXECUTE_TIMEOUT_US, rc
-        );
-    } else {
-        LOG_INFO_V0(
-            "aclrtSetOpExecuteTimeOutV2: requested=%llu us, actual=%llu us",
-            (unsigned long long)PLATFORM_OP_EXECUTE_TIMEOUT_US, (unsigned long long)actual_timeout
-        );
-    }
-}
-
-int DeviceRunner::ensure_binaries_loaded() {
-    // Check if already loaded (binaries owned by the runner via set_executors).
-    if (binaries_loaded_) {
-        return 0;
-    }
-
-    // Device must be set first
-    if (stream_aicpu_ == nullptr) {
-        LOG_ERROR("Device not set before loading binaries");
-        return -1;
-    }
-
-    if (dispatcher_so_binary_.empty()) {
-        LOG_ERROR(
-            "DeviceRunner: dispatcher SO bytes not provided; pass dispatcher_path through ChipWorker.init "
-            "(RuntimeBinaries.dispatcher_path)"
-        );
-        return -1;
-    }
-
-    // One-shot bootstrap: libaicpu_extend_kernels invokes our dispatcher,
-    // which writes the runtime AICPU SO bytes to simpler_inner_<fp>_<device_id>.so in
-    // the device-side preinstall path. The dispatcher SO itself is never
-    // persisted. Subsequent per-task AICPU launches resolve symbols via
-    // rtsBinaryLoadFromFile + rtsFuncGetByName + rtsLaunchCpuKernel
-    // directly against that preinstall file.
-    int rc = load_aicpu_op_.BootstrapDispatcher(
-        dispatcher_so_binary_.data(), dispatcher_so_binary_.size(), aicpu_so_binary_.data(), aicpu_so_binary_.size(),
-        stream_aicpu_, device_id_
-    );
-    if (rc != 0) {
-        LOG_ERROR("LoadAicpuOp::BootstrapDispatcher failed: %d", rc);
-        return rc;
-    }
-    LOG_INFO_V2("DeviceRunner: inner SO uploaded to preinstall via dispatcher bootstrap");
-
-    rc = load_aicpu_op_.Init();
-    if (rc != 0) {
-        LOG_ERROR("LoadAicpuOp::Init failed: %d", rc);
-        return rc;
-    }
-    LOG_INFO_V2("DeviceRunner: inner SO registered (simpler_aicpu_init/exec handles ready)");
-
-    // H2D the per-task DeviceArgs struct itself. device_args_.aicpu_so_bin/len
-    // stay zero — our own per-task AICPU code (launched via rtsLaunchCpuKernel
-    // against the cached rtFuncHandle on LoadAicpuOp) never reads them, and
-    // the dispatcher-bootstrap KernelArgs (KERNEL_TYPE_AICPU_KFC) builds its
-    // own DeviceArgs view inside BootstrapDispatcher rather than reading
-    // ours. The "load-bearing on a5" finding documented prior to #864/#870
-    // no longer reproduces against current HEAD — see PR removing
-    // AicpuSoInfo (CI on both archs green).
-    rc = kernel_args_.init_device_args(device_args_, mem_alloc_);
-    if (rc != 0) {
-        LOG_ERROR("init_device_args failed: %d", rc);
-        return rc;
-    }
-
-    // Release host bytes — bootstrap is done. Per-task launches go through
-    // the cached rtFuncHandle owned by LoadAicpuOp; dispatcher SO bytes are
-    // never referenced again; the aicpu kernel SO's host buffer is no longer
-    // needed either (we used to H2D it through AicpuSoInfo as a CANN-internal
-    // bookkeeping workaround; that's gone).
-    dispatcher_so_binary_.clear();
-    dispatcher_so_binary_.shrink_to_fit();
-    aicpu_so_binary_.clear();
-    aicpu_so_binary_.shrink_to_fit();
-
-    binaries_loaded_ = true;
-    LOG_INFO_V0("DeviceRunner: binaries loaded");
-    return 0;
-}
-
-int DeviceRunner::query_max_block_dim(rtStream_t stream, uint32_t *out_cube, uint32_t *out_vector) {
-    uint32_t cube_limit = 0, vector_limit = 0;
-    bool got_limits = (aclrtGetStreamResLimit(stream, ACL_RT_DEV_RES_CUBE_CORE, &cube_limit) == ACL_ERROR_NONE) &&
-                      (aclrtGetStreamResLimit(stream, ACL_RT_DEV_RES_VECTOR_CORE, &vector_limit) == ACL_ERROR_NONE) &&
-                      cube_limit > 0 && vector_limit > 0;
-    if (out_cube != nullptr) *out_cube = got_limits ? cube_limit : 0;
-    if (out_vector != nullptr) *out_vector = got_limits ? vector_limit : 0;
-    if (got_limits) {
-        // Cap by PLATFORM_MAX_BLOCKDIM as well: runtime handshake/scheduler
-        // arrays are statically sized to RUNTIME_MAX_WORKER (= PLATFORM_MAX_BLOCKDIM
-        // * PLATFORM_CORES_PER_BLOCKDIM), so even if ACL reports more cores
-        // than the platform cap we must not exceed it.
-        int from_stream = static_cast<int>(
-            std::min(cube_limit / PLATFORM_AIC_CORES_PER_BLOCKDIM, vector_limit / PLATFORM_AIV_CORES_PER_BLOCKDIM)
-        );
-        return std::min(from_stream, PLATFORM_MAX_BLOCKDIM);
-    }
-    return PLATFORM_MAX_BLOCKDIM;
-}
-
-int DeviceRunner::validate_block_dim(rtStream_t stream, int block_dim) {
-    if (block_dim < 1) {
-        LOG_ERROR("block_dim (%d) must be >= 1", block_dim);
-        return -1;
-    }
-    uint32_t cube_limit = 0, vector_limit = 0;
-    int max_bd = query_max_block_dim(stream, &cube_limit, &vector_limit);
-    if (block_dim > max_bd) {
-        if (cube_limit > 0 && vector_limit > 0) {
-            LOG_ERROR(
-                "block_dim (%d) exceeds available cores (max_block_dim=%d, cube=%u, vector=%u)", block_dim, max_bd,
-                cube_limit, vector_limit
-            );
-        } else {
-            LOG_ERROR(
-                "aclrtGetStreamResLimit unavailable; block_dim (%d) exceeds static cap PLATFORM_MAX_BLOCKDIM (%d)",
-                block_dim, PLATFORM_MAX_BLOCKDIM
-            );
-        }
-        return -1;
-    }
-    return 0;
-}
+// `create_thread`, `attach_current_thread`, `configure_aicore_op_timeout`,
+// `ensure_device_initialized`, `ensure_binaries_loaded`, `query_max_block_dim`,
+// and `validate_block_dim` live on `DeviceRunnerBase` — see
+// `src/common/platform/onboard/host/device_runner_base.cpp`.
 
 int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
     if (launch_aicpu_num < 1 || launch_aicpu_num > PLATFORM_MAX_AICPU_THREADS) {
@@ -617,24 +414,7 @@ int DeviceRunner::run(Runtime &runtime, int block_dim, int launch_aicpu_num) {
     return 0;
 }
 
-void DeviceRunner::print_handshake_results() {
-    if (stream_aicpu_ == nullptr || worker_count_ == 0 || kernel_args_.args.runtime_args == nullptr) {
-        return;
-    }
-
-    // Allocate temporary buffer to read handshake data from device
-    std::vector<Handshake> workers(worker_count_);
-    size_t total_size = sizeof(Handshake) * worker_count_;
-    rtMemcpy(workers.data(), total_size, kernel_args_.args.runtime_args->workers, total_size, RT_MEMCPY_DEVICE_TO_HOST);
-
-    LOG_DEBUG("Handshake results for %d cores:", worker_count_);
-    for (int i = 0; i < worker_count_; i++) {
-        LOG_DEBUG(
-            "  Core %d: aicore_done=%d aicpu_ready=%d task=%d", i, workers[i].aicore_done, workers[i].aicpu_ready,
-            workers[i].task
-        );
-    }
-}
+// `print_handshake_results` lives on `DeviceRunnerBase`.
 
 int DeviceRunner::prepare_orch_so(Runtime &runtime) {
     // Prepared-callable flow only: bytes were H2D'd at prepare_callable
